+++ example/lib/main.dart
@@ -1,68 +1,36 @@
+import 'package:example/src/screen/home_page.dart';
+import 'package:example/src/screen/login_page.dart';
+import 'package:example/src/util/route_generator.dart';
+import 'package:example/src/util/storage_key.dart';
+import 'package:example/src/util/theme.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttered/fluttered.dart';
-import 'package:example/src/util/theme.dart';
 
-void main() {
-  flutteredConfig(theme: ThemeConfig(availableThemes, defaultTheme));
+void main() async {
+  await flutteredConfig();
   runApp(MyApp());
 }
 
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return ThemeManager(
-      builder: (context, theme) => MaterialApp(
-        title: 'Flutter Demo',
-        theme: theme,
-        home: MyHomePage(title: 'Flutter Demo Home Page'),
-      ),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
-
-  final String title;
-
-  @override
-  _MyHomePageState createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  int _counter = 0;
-
-  void _incrementCounter(BuildContext context) {
-    setState(() {
-      _counter++;
-    });
-    ThemeManager.of(context).setTheme('theme2');
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.title),
-      ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(
-              'You have pushed the button this many times:',
-            ),
-            Text(
-              '$_counter',
-              style: Theme.of(context).textTheme.display1,
-            ),
-          ],
+    return LifecycleManager(
+      onPaused: () => print('PAUSED'),
+      onResumed: () => print('RESUMED'),
+      child: ThemeManager(
+        availableThemes: availableThemes,
+        defaultTheme: defaultTheme,
+        builder: (context, theme) => MaterialApp(
+          title: 'Fluttered',
+          theme: theme,
+          onGenerateRoute: RouteGenerator.generateRoute,
+          home: LandingPageManager(
+            homeScreen: HomePage(title: 'Home'),
+            requireLogin: true,
+            loggedInKey: loggedInKey,
+            loginScreen: LoginPage(),
+          ),
         ),
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () => _incrementCounter(context),
-        tooltip: 'Increment',
-        child: Icon(Icons.add),
       ),
     );
   }