+++ lib/fluttered.dart
@@ -1,5 +1,6 @@
 library fluttered;
 
-export 'src/widget/widget.dart';
-export 'src/util/util.dart';
-export 'src/config/config.dart';
+export 'src/config/index.dart';
+export 'src/global/index.dart';
+export 'src/service/index.dart';
+export 'src/widget/index.dart';