+++ lib/src/config/fluttered_config.dart
@@ -1,8 +1,8 @@
-import 'package:fluttered/src/config/theme_config.dart';
+import 'package:fluttered/src/global/public_instance.dart';
+import 'package:fluttered/src/service/shared_prefs_service.dart';
 
 /// Set Global configurations for fluttered.
-void flutteredConfig({
-  ThemeConfig theme,
-}) {
-  themeConfig = theme;
+Future<void> flutteredConfig() async {
+  // Manadatory
+  sharedPrefsServiceInstance = await SharedPrefsService.instance();
 }