+++ lib/src/config/index.dart
@@ -1,2 +1,1 @@
 export 'fluttered_config.dart';
-export 'theme_config.dart';