+++ lib/src/global/public_instance.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/widgets.dart';
+import 'package:fluttered/src/service/shared_prefs_service.dart';
+
+// INSTANCES
+
+/// Instance of [SharedPrefsService].
+SharedPrefsService sharedPrefsServiceInstance;
+
+// STREAMS
+
+/// Stream of application lifecycle states.
+Stream<AppLifecycleState> lifecycleStateStream;