+++ lib/src/service/index.dart
@@ -1,0 +1,1 @@
+export 'shared_prefs_service.dart';