+++ lib/src/widget/index.dart
@@ -1,3 +1,5 @@
 export 'stateful_wrapper.dart';
 export 'pvc.dart';
 export 'theme_manager.dart';
+export 'lifecycle_manager.dart';
+export 'landing_page_manager.dart';